refactor: move mock functions to class in recipe service test
